RIDC using Oracle Java Stored Procedures
I will explain how use a table with content in a BLOB column and check
them into the Content Server. We will use two approaches, one to save
the file to the disk and then checkin, and the other to keep the file
in memory and checkin. Note that the in memory solution should not be
used with large files.