WebCenter Content in the Enterprise
The different WebCenter Content modules and how they fit into the
enterprise. Since then, the product branding has changed, so made some
updates and posted it here.

For the sake of simplicity, only Oracle WebCenter Content products are
shown here. Oracle Forms Recognition is not part of the suite but is
shown for completeness. There are out-of-the-box integrations/solutions
with other Oracle Fusion Middleware products like WebCenter Portal and
SOA/BPM Suite which are not shown here.